# Optimize contextgraph batch send algorithm