This function attempts to split the request size to be the largest
possible request that is smaller than the request size limit. It did
this by linearly checking size(list[:n]), size(list[:n-1])... This ends
up being extremely slow, because proto.size is not a very cheap
operation for such large objects. In the case of racetests, this test
was taking over 5 minutes sometimes.

This modifies the algorithm to do a binary search for the optimal
request; behavior should be the same.

With this change:
BenchmarkSendBatch/Size2-8                500000              3042 ns/op
BenchmarkSendBatch/Size200-8               10000            135250 ns/op
BenchmarkSendBatch/Size20000-8                30          52180768 ns/op

Without this change
BenchmarkSendBatch/Size2-8                500000              3082 ns/op
BenchmarkSendBatch/Size200-8               10000            140837 ns/op
BenchmarkSendBatch/Size20000-8                 1        7999998085 ns/op

This represents a 150x improvement on large request sizes and a
negligible change for small requests

Alternatively, if we don't care about the performance of this function,
we can disable it in the racetests and leave this code as is.

[x] Policies and Telemetry