@istio-testing 2019-07-22
@howardjohn: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
e2e-bookInfoTests-envoyv2-v1alpha3-master | b4ad62b59070ab2abc6b4f5cc8489a7929b7f0d4 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/15689/e2e-bookInfoTests-envoyv2-v1alpha3-master/4049) | `/test e2e-bookInfoTests-envoyv2-v1alpha3-master`
integ-mixer-k8s-presubmit-tests-master | b4ad62b59070ab2abc6b4f5cc8489a7929b7f0d4 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/15689/integ-mixer-k8s-presubmit-tests-master/3035) | `/test integ-mixer-k8s-presubmit-tests-master`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->