@bianpengyuan 2019-07-22
Would it be nicer to calculate size of each `asst` and accumulate at append time, then create a new empty req if the current one exceed size limit? With this, req will already be well split at sending time so no search is needed.