@howardjohn 2019-07-22
> Would it be nicer to calculate size of each asst and accumulate at append time, then create a new empty req if the current one exceed size limit? With this, req will already be well split at sending time so no search is needed.

The thing I wasn't sure about is if the sum of the size of each `asst` is the same as the size of the full list of asst. I suppose since we have a buffer on the max size (550kb is the max in the file but a comment claims 600kb is the real max) then it could work