@bianpengyuan 2019-07-22
Yeah it is a rough estimate and with some head room reserved on size limit it should be safe.