@istio-testing 2019-07-22
@howardjohn: The following test **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-mixer-k8s-presubmit-tests-master | 99c770957eed5f43b448aa6df2ac368dbb23e208 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/15689/integ-mixer-k8s-presubmit-tests-master/3071) | `/test integ-mixer-k8s-presubmit-tests-master`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->